requirements.txt
streamlit
google-cloud-aiplatform
python-dotenv
flask
